move clis to separate module
